+++ R/rix_helpers.R
@@ -420,6 +420,7 @@ generate_shell <- function(
   flag_git_archive,
   flag_rpkgs,
   flag_tex_pkgs,
+  py_conf,
   flag_py_conf,
   flag_local_r_pkgs,
   flag_wrapper,
@@ -436,7 +437,7 @@ generate_shell <- function(
   };",
     generate_locale_archive(detect_os()),
     generate_locale_variables(),
-    generate_set_reticulate(flag_py_conf),
+    generate_set_reticulate(py_conf, flag_py_conf),
     flag_git_archive,
     flag_rpkgs,
     flag_tex_pkgs,
@@ -472,10 +473,14 @@ remove_empty_lines <- function(default.nix) {
 
 #' generate_set_reticulate Helper to set path to reticulate
 #' @noRd
-generate_set_reticulate <- function(flag_py_conf) {
+generate_set_reticulate <- function(py_conf, flag_py_conf) {
   if (flag_py_conf == "") {
     ""
   } else {
-    paste0('RETICULATE_PYTHON = "${pkgs.python312}/bin/python";\n')
+    py_version <- paste0(
+      "python",
+      gsub("\\.", "", py_conf$py_version)
+    )
+    paste0('RETICULATE_PYTHON = "${pkgs.', py_version, '}/bin/python";\n')
   }
 }